refactor(useValidation): rename extension to JS instead of JSX (no JSX in file)
