+++ src/components/checkbox/inline-checkbox.jsx
@@ -10,7 +10,7 @@ import PropTypes from 'prop-types';
 import Icon from '../icon/icon.jsx';
 import HelpIcon from '../help-icon/help-icon.jsx';
 import cautionSvg from '../../svgs/caution.svg';
-import useValidation from '../../hooks/use-validation.jsx';
+import useValidation from '../../hooks/use-validation.js';
 import styles from './checkbox.css';
 
 const Checkbox = forwardRef(function Checkbox(
