+++ src/components/textarea/textarea.jsx
@@ -7,7 +7,7 @@ import React, {
 import FormControl from '../form-control/form-control.jsx';
 import Icons from '../control/icons.jsx';
 import useMountedLayoutEffect from '../../hooks/use-mounted-layout-effect.js';
-import useValidation from '../../hooks/use-validation.jsx';
+import useValidation from '../../hooks/use-validation.js';
 import styles from './textarea.css';
 
 function getClassName(className, validationMessage) {
